Created JQuery plugin and Component Enhancement.
Check below enhancement points:

1. Added 'data-target' attribute in the Sidebar to provide the facility
to add multiple sidebars. The value of the 'data-target' attribute will
be the id of the 'aside' to be called. For example, If
data-target=''basicSidebar" is added in the anchor tag - <a
href="javascript:void(0);" data-target="basicSidebar></a>,  the same
id="basicSidebar” needs to be added in the sidebar container div <aside
id="basicSidebar" class="pmd-sidebar">.

2. To create different types of sidebars, we have removed sidebar
configuration classes '.pmd-sidebar-slide-push',
'.pmd-sidebar-left-fixed', '.pmd-sidebar-left' and '.pmd-sidebar-open'
from '.pmd-sidebar' and replaced them with data attributes
'data-position=”slidepush/fixed”, data-placement="left/right" and
'is-open="true/false" respectively.

3. Added 'is-open-width="1000"' data attribute in the anchor tag of the
sidebar using which the width of the sidebar can be customized.